+++ python/baseline/pytorch/lm/model.py
@@ -34,7 +34,7 @@ class LanguageModelBase(nn.Module, LanguageModel):
 
         y = batch_dict.get('y')
         if y is not None:
-            y = torch.from_numpy(y)
+            y = torch.from_numpy(y).to(torch.long)
             if self.gpu is not None:
                 y = y.cuda()
             example_dict['y'] = y
